Artificial-Intelligence-Conducting-Sentiment-Analysis-on-IMDB-Reviews
This project focuses on an artificial intelligence system that analyzes movie reviews on IMDB. Artificial intelligence algorithms are trained to understand the emotional content in these reviews using natural language processing (NLP) techniques. The emotional tones in the reviews are categorized into different categories such as positive, negative, or neutral, providing an overall emotional feedback about the movie. Such sentiment analysis can be utilized to provide users with a deeper understanding of a specific movie and assist audiences in making informed decisions about the film.